Move VelocitySmoothing to new motion_planning library
