+++ src/lib/CMakeLists.txt
@@ -54,6 +54,7 @@ add_subdirectory(led)
 add_subdirectory(mathlib)
 add_subdirectory(mixer)
 add_subdirectory(mixer_module)
+add_subdirectory(motion_planning)
 add_subdirectory(output_limit)
 add_subdirectory(perf)
 add_subdirectory(pid)
