+++ src/lib/flight_tasks/tasks/Utility/ManualVelocitySmoothingXY.hpp
@@ -38,7 +38,7 @@
 
 #pragma once
 
-#include "VelocitySmoothing.hpp"
+#include <motion_planning/VelocitySmoothing.hpp>
 
 #include <matrix/matrix/math.hpp>
