+++ src/lib/motion_planning/CMakeLists.txt
@@ -31,16 +31,10 @@
 #
 ############################################################################
 
-px4_add_library(FlightTaskUtility
-	StraightLine.cpp
-	ManualVelocitySmoothingXY.cpp
-	ManualVelocitySmoothingZ.cpp
-	Sticks.cpp
-	StickAccelerationXY.cpp
-	StickYaw.cpp
+px4_add_library(motion_planning
+	VelocitySmoothing.cpp
 )
 
-target_link_libraries(FlightTaskUtility PUBLIC FlightTask hysteresis bezier SlewRate motion_planning)
-target_include_directories(FlightTaskUtility PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(motion_planning PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
-px4_add_unit_gtest(SRC ManualVelocitySmoothingXYTest.cpp LINKLIBS FlightTaskUtility)
+px4_add_unit_gtest(SRC VelocitySmoothingTest.cpp LINKLIBS motion_planning)
